requirements.txt
# Core
playwright==1.48.0
python-dotenv==1.0.1

# UI
streamlit==1.40.0

# Export
python-docx==1.1.2

# Scheduler
apscheduler==3.10.4

# Database
sqlalchemy==2.0.35

# Email
aiosmtplib==3.0.2

# Utilities
pydantic==2.9.2
python-dateutil==2.9.0
pytz==2024.2

# Async
aiofiles==24.1.0

# HTTP Client (for OpenAI API)
httpx==0.27.0

# Alternativa Crawl4AI (opcional)
# crawl4ai==0.3.0

# Testes
pytest==8.3.3
pytest-asyncio==0.24.0
